backend/admin-service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
httpx
python-dotenv
loguru
cryptography